Functional Game
Today I was able to run the game for the project, however there are
some slight changes that need to be made. My teammate on the otherhand
was working on the map for our game today. I feel that we have been very
productive and effective in lesson today.

Estimate for remaining time to finish assignment: quite a few lessons