Finished sprite_directionality and itemmixin and started puzzle
There was a lot of work completed today. I managed to finish the
sprite_directionality and also the InventoryItemMixin. I also started
working on the puzzle.py and also research into creating the puzzles.
Most of the foundation of the game is completed now, however my
teammate who made the game map is experiencing some difficulties
importing the map.

Estimate for remaining time to finish assignment: a few classes